fix/move files to bot/cor & removed very old code2
