+++ bot/core/shard.js
